efi/libstub: Use const qualified device path protocol pointers
The EFI APIs that deal with device path protocol instances never modify
them, but given that these APIs do not use const qualified pointers at
all, but instead uses IN and OUT modifiers that are #define'd to empty
tokens, we cannot currently pass a device path protocol pointer that
refers to a read-only object.

So add the const qualification where appropriate, and drop a cast that
is no longer necessary.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>