io_uring/uring_cmd: push IRQ based completions through task_work
This is similar to what we do on the non-passthrough read/write side,
and helps take advantage of the completion batching we can do when we
post CQEs via task_work. On top of that, this avoids a uring_lock
grab/drop for every completion.

In the normal peak IRQ based testing, this increases performance in
my testing from ~75M to ~77M IOPS, or an increase of 2-3%.

Signed-off-by: Jens Axboe <axboe@kernel.dk>