btrfs: remove the sync_writers field in struct btrfs_inode
The sync_writers field communicates if an inode has outstanding
synchronous writeback.  The better way is to look at the REQ_SYNC
flag in the bio, which is set by the writeback code for WB_SYNC_ALL
writeback and can communicate the need for sync writeback without
an inode field.

Reviewed-by: Chris Mason <clm@fb.com>
Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: David Sterba <dsterba@suse.com>