iio: adc: meson: add support for the GXLX SoC - WiP
The SARADC IP on the GXLX SoC itself is identical to the one found on
GXL SoCs. However, GXLX SoCs require poking the first three bits in the
MESON_SAR_ADC_REG12 register to get the three MPLL clocks (used as clock
generators for the audio frequencies) to work.

WiP: the purpose of these three bits needs to be clarified

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>