regulator: pwm-regulator: Make assumptions about disabled PWM consistent
Generally it's not known how a disabled PWM behaves. However if the
bootloader hands over a disabled PWM that drives a regulator and it's
claimed the regulator is enabled, then the most likely assumption is
that the PWM emits the inactive level. This is represented by duty_cycle
= 0 even for .polarity == PWM_POLARITY_INVERSED.

Change the implementation to always use duty_cycle = 0, regardless of
the polarity. Also update the code so it keeps a copy of the pwm_state
around. Both of these changes result in the fact that the logic from
pwm_regulator_init_boot_on() is much less complex and can be simplified
to a point where the whole function is not needed anymore.

Suggested-by: Uwe Kleine-König <u.kleine-koenig@baylibre.com>
Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>