net: phy: meson-gxl: Add support for banked read/write without test mode
The existing initialization within the meson-gxl PHY driver requires the
TEST_MODE bit to be set. This however is not the case for all read/write
operations as WOL (wake-on-lan) support for example has to be configured
without the TEST_MODE bit set.
Update the banked read/write code so callers can opt in/out of test
mode.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>